requirements.txt
amqp==2.1.4
appdirs==1.4.2
appnope==0.1.0
asgi-redis==1.0.0
asgiref==1.0.0
attrs==16.3.0
autobahn==0.17.2
Automat==0.5.0
backports.shutil-get-terminal-size==1.0.0
billiard==3.5.0.2
celery==4.0.2
channels==1.0.3
constantly==15.1.0
daphne==1.0.3
decorator==4.0.11
Django==1.10.6
enum34==1.1.6
google-api-python-client==1.6.2
httplib2==0.10.3
incremental==16.10.1
ipython==5.2.2
ipython-genutils==0.1.0
kombu==4.0.2
msgpack-python==0.4.8
oauth2client==4.0.0
packaging==16.8
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
ptyprocess==0.5.1
pyasn1==0.2.2
pyasn1-modules==0.0.8
Pygments==2.2.0
pypandoc==1.3.3
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
redis==2.10.5
rsa==3.4.2
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
Twisted==17.1.0
txaio==2.6.1
uritemplate==3.0.0
vine==1.1.3
wcwidth==0.1.7
zope.interface==4.3.3
